// src/app/features/mastertables/tblHSNMaster/models/tblHSNMaster-Update.model.ts
export interface TblHSNMasterUpdate {
	fldId: number;
	fldFKHSNCategory: number;
	fldItemNames: string;
	fldSalesCode: string;
	fldJobworkCode: string;
	fldGSTSalesPercentage: number;
	fldGSTJobworkPercentage: number;
	fldIsActive: boolean;
	fldCreatedBy: number;
	fldCreatedDt: Date;
	fldModifiedBy: number;
	fldModifiedDt: Date;
}